Fix : 여러가지 수정..
