+++ src/Pages/shared/SharedPage.js
@@ -1,6 +1,7 @@
 import useAsync from "../../Hooks/useAsync";
 import getSample from "../../api";
-import Folder from "../../components/Folder/Folder";
+import Folder from "../../components/Folder";
+
 import Footer from "../../components/Footer/Footer";
 import Nav from "../../components/common/Nav";
