--- src/components/CardList/CardListStyle.js
@@ -1,18 +1,0 @@
-import styled from 'styled-components'
-import { device } from '../../style/device'
-
-export const CardListContainer = styled.div`
-  display: grid;
-
-  gap: 20px;
-  grid-template-columns: repeat(3, 1fr);
-
-  @media ${device.tablet} {
-    grid-template-columns: repeat(2, 1fr);
-  }
-
-  @media ${device.mobile} {
-    place-items: center;
-    grid-template-columns: repeat(1, 1fr);
-  }
-`
