+++ src/components/FolderAndLink.jsx
@@ -2,20 +2,16 @@ import { useEffect, useState } from "react";
 
 import { getFolders, getLinksByFolderID } from "../api";
 import FolderList from "./FolderList/FolderList";
-import CardList from "./CardList/CardList";
+
 import useAsync from "../Hooks/useAsync";
 
 import styled from "styled-components";
 import { device } from "../style/device";
-import { useSearchParams } from "react-router-dom";
-
+import CardList from "./CardList";
 
 function FolderAndLink() {
   const [selectedFolderId, setSelectedFolderId] = useState(undefined);
-  // const [searchParams] = useSearchParams();
-  // const folderId = searchParams.get("folderId");
 
-  
   const DEFAULT_FOLDER = 1;
   // 폴더
